rm data-directory, and move mapper to delivery-package
